Move classes to appropriate packages and reduce visibility to `package-private`
