+++ rdbms/src/main/java/io/spine/server/storage/jdbc/aggregate/InsertLifecycleFlagsQuery.java
@@ -22,7 +22,7 @@ package io.spine.server.storage.jdbc.aggregate;
 
 import com.querydsl.core.dml.StoreClause;
 import io.spine.server.entity.LifecycleFlags;
-import io.spine.server.storage.jdbc.IdColumn;
+import io.spine.server.storage.jdbc.query.IdColumn;
 import io.spine.server.storage.jdbc.query.AbstractStoreQuery;
 import io.spine.server.storage.jdbc.query.Parameter;
 import io.spine.server.storage.jdbc.query.Parameters;
