+++ rdbms/src/main/java/io/spine/server/storage/jdbc/aggregate/UpdateLifecycleFlagsQuery.java
@@ -23,7 +23,7 @@ package io.spine.server.storage.jdbc.aggregate;
 import com.querydsl.core.dml.StoreClause;
 import com.querydsl.core.types.dsl.PathBuilder;
 import io.spine.server.entity.LifecycleFlags;
-import io.spine.server.storage.jdbc.IdColumn;
+import io.spine.server.storage.jdbc.query.IdColumn;
 import io.spine.server.storage.jdbc.query.AbstractStoreQuery;
 import io.spine.server.storage.jdbc.query.Parameter;
 import io.spine.server.storage.jdbc.query.Parameters;
