+++ rdbms/src/main/java/io/spine/server/storage/jdbc/aggregate/WriteEventCountQuery.java
@@ -20,7 +20,7 @@
 
 package io.spine.server.storage.jdbc.aggregate;
 
-import io.spine.server.storage.jdbc.IdColumn;
+import io.spine.server.storage.jdbc.query.IdColumn;
 import io.spine.server.storage.jdbc.query.AbstractStoreQuery;
 import io.spine.server.storage.jdbc.query.Parameter;
 import io.spine.server.storage.jdbc.query.Parameters;
