+++ rdbms/src/test/java/io/spine/server/storage/jdbc/aggregate/CloseablesShould.java
@@ -18,13 +18,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.storage.jdbc;
+package io.spine.server.storage.jdbc.aggregate;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.google.common.testing.NullPointerTester;
-import io.spine.server.storage.jdbc.Closeables;
-import io.spine.server.storage.jdbc.MultipleExceptionsOnClose;
 import org.junit.Test;
 
 import java.util.Collection;
